requirements.txt
mysql-connector-python
python-dotenv
fastapi
uvicorn
starlette